Move extensions.
