+++ coil-video/api/coil-video.api
@@ -1,11 +1,3 @@
-public final class coil/extension/Videos {
-	public static final fun videoFrameMicros (Lcoil/request/ImageRequest$Builder;J)Lcoil/request/ImageRequest$Builder;
-	public static final fun videoFrameMicros (Lcoil/request/Parameters;)Ljava/lang/Long;
-	public static final fun videoFrameMillis (Lcoil/request/ImageRequest$Builder;J)Lcoil/request/ImageRequest$Builder;
-	public static final fun videoFrameOption (Lcoil/request/ImageRequest$Builder;I)Lcoil/request/ImageRequest$Builder;
-	public static final fun videoFrameOption (Lcoil/request/Parameters;)Ljava/lang/Integer;
-}
-
 public abstract class coil/fetch/VideoFrameFetcher : coil/fetch/Fetcher {
 	public static final field Companion Lcoil/fetch/VideoFrameFetcher$Companion;
 	public static final field VIDEO_FRAME_MICROS_KEY Ljava/lang/String;
@@ -35,5 +27,13 @@ public final class coil/fetch/VideoFrameUriFetcher : coil/fetch/VideoFrameFetche
 	public fun key (Landroid/net/Uri;)Ljava/lang/String;
 	public synthetic fun key (Ljava/lang/Object;)Ljava/lang/String;
 	public synthetic fun setDataSource (Landroid/media/MediaMetadataRetriever;Ljava/lang/Object;)V
+}
+
+public final class coil/request/Videos {
+	public static final fun videoFrameMicros (Lcoil/request/ImageRequest$Builder;J)Lcoil/request/ImageRequest$Builder;
+	public static final fun videoFrameMicros (Lcoil/request/Parameters;)Ljava/lang/Long;
+	public static final fun videoFrameMillis (Lcoil/request/ImageRequest$Builder;J)Lcoil/request/ImageRequest$Builder;
+	public static final fun videoFrameOption (Lcoil/request/ImageRequest$Builder;I)Lcoil/request/ImageRequest$Builder;
+	public static final fun videoFrameOption (Lcoil/request/Parameters;)Ljava/lang/Integer;
 }
